+++ changelog.md
@@ -1,3 +1,9 @@
+v0.8.5
+------
+
+- Javascript is loaded via the `WebView.executeScript()` method, closes #183.
+- Added Linux build, closes #181.
+
 v0.8.4
 ------